func/requirements.txt
azure-functions
nest-asyncio
fastapi
strawberry-graphql[fastapi]
